+++ tests/test_options.py
@@ -1,3 +1,9 @@
+try:
+    from urllib.parse import quote
+except ImportError:
+    # Python 2.7 compatibility
+    from urllib import quote
+
 import pytest
 from marshmallow import validate, ValidationError
 
@@ -76,13 +82,15 @@ class TestInflection:
 
         # valid
         data = unpack(schema.load(make_serialized_author({
-            'first-name': 'Nevets', 'last-name': 'Longoria'})))
+            'first-name': 'Nevets', 'last-name': 'Longoria',
+        })))
         assert data['first_name'] == 'Nevets'
 
     def test_load_with_inflection_and_load_from_override(self):
         schema = AuthorSchemaWithOverrideInflection()
         data = unpack(schema.load(make_serialized_author({
-            'firstName': 'Steve', 'last-name': 'Loria'})))
+            'firstName': 'Steve', 'last-name': 'Loria',
+        })))
         assert data['first_name'] == 'Steve'
         assert data['last_name'] == 'Loria'
 
@@ -103,7 +111,7 @@ class TestInflection:
             post_comments = fields.Relationship(
                 'http://test.test/posts/{id}/comments/',
                 related_url_kwargs={'id': '<id>'},
-                attribute='comments'
+                attribute='comments',
             )
 
             class Meta:
@@ -129,6 +137,14 @@ class AuthorAutoSelfLinkSchema(Schema):
         type_ = 'people'
         self_url = '/authors/{id}'
         self_url_kwargs = {'id': '<id>'}
+        self_url_many = '/authors/'
+
+class AuthorAutoSelfLinkFirstLastSchema(AuthorAutoSelfLinkSchema):
+
+    class Meta:
+        type_ = 'people'
+        self_url = '/authors/{first_name} {last_name}'
+        self_url_kwargs = {'first_name': '<first_name>', 'last_name': '<last_name>'}
         self_url_many = '/authors/'
 
 
@@ -168,3 +184,11 @@ class TestAutoSelfUrls:
         assert first['type'] == 'comments'
 
         assert 'links' not in data
+
+    def test_self_link_quoted(self, author):
+        data = unpack(AuthorAutoSelfLinkFirstLastSchema().dump(author))
+        assert 'links' in data
+        assert data['links']['self'] == quote('/authors/{} {}'.format(
+            author.first_name,
+            author.last_name,
+        ))